% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoD-Test.R
\name{DoD.boot.samp}
\alias{DoD.boot.samp}
\title{Approximation of the limit distribution under the null hypothesis}
\usage{
DoD.boot.samp(
  input,
  beta = 0.01,
  p = 2,
  boot.m.out.of.n = dim(input)[1],
  boot.rep = 1000,
  force.coordinates = F,
  method = "euclidean"
)
}
\arguments{
\item{input}{either a n x d numeric coordinate matrix (resp. data frame) or a n x n distance matrix.}

\item{beta}{the trimming parameter of the DoD-statistic. It can attain values in [0,1/2).}

\item{p}{the order of the DoD-statistic. By definition p>= 1.}

\item{boot.m.out.of.n}{the bootstrap sample size.}

\item{boot.rep}{the number of bootstrap repetitions.}

\item{force.coordinates}{boolean that indicates whether the input is forced to be interpreted as coordinate matrix.}

\item{method}{the distance measure to be used in case the input is considered as coordinate matrices. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}
}
\value{
The function returns a numeric vector that contains the specified number (given by \code{boot.rep}) of m-out-of-n bootstrap realizations (m=\code{boot.m.out.of.n }) of the resampling
scheme proposed in Section 3 of \insertCite{DoD2021;textual}{DoD} in order to approximate the limit distribution of the DoD-Statistic under the null hypothesis.
}
\description{
Calculate a bootstrap approximation of the limit distribution of the DoD-statistic under the null
hypothesis based on a sample of size n (see Section 3 of \insertCite{DoD2021;textual}{DoD} for additional information).
}
\details{
In order to apply the DoD-Test the limit distribution of the DoD-Statistic under the null
hypothesis
 
\strong{H:}  \emph{"The distributions of (interpoint) distances of both inputs are equal."}
 
is required. This function implements the resampling scheme proposed in Section 3 of \insertCite{DoD2021;textual}{DoD} to approximate this limit distribution.  
 
If the input is a quadratic numeric matrix (and \code{force.coordinates}==F), it will be interpreted as distance matrix. Otherwise, it is considered as coordinate matrix
and the interpoint distances are calculated with \code{dist} and the method specified.
}
\examples{
n= 100

square = data.frame("x"=numeric(n),"y"=numeric(n))
square$x = runif(n,0,1)
square$y = runif(n,0,1)


DoD.boot.samp(square)

}
\references{
{
  \insertAllCited{}
}
}
