% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoD-Test.R
\name{DoD.stat}
\alias{DoD.stat}
\title{The DoD-Statistic}
\usage{
DoD.stat(
  input.1,
  input.2,
  beta = 0.01,
  p = 2,
  force.coordinates = F,
  method = "euclidean"
)
}
\arguments{
\item{input.1}{either a n x d numeric coordinate matrix (resp. data frame) or a n x n distance matrix.}

\item{input.2}{either a m x d numeric coordinate matrix (resp. data frame) or a m x m distance matrix.}

\item{beta}{the trimming parameter of the DoD-statistic. It can attain values in [0,1/2).}

\item{p}{the order of the DoD-statistic. By definition p>= 1.}

\item{force.coordinates}{boolean that indicates whether the input is forced to be interpreted as coordinate matrix.}

\item{method}{the distance measure to be used in case the input is considered as coordinate matrices. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}
}
\value{
The function returns the value of the DoD-statistic (defined in \insertCite{DoD2021;textual}{DoD}).
}
\description{
Calculate the (unscaled) DoD-statistic proposed in \insertCite{DoD2021;textual}{DoD} for two samples of size n and m.
}
\details{
The DoD-Statistic is defined as the (trimmed) empirical Kantorovich distance of order p between the distribution of distances of the input
samples \code{input1} and \code{input2} (for more details see \insertCite{DoD2021;textual}{DoD}).
 
If both inputs are quadratic numeric matrices (and \code{force.coordinates}==F), they will be interpreted as distance matrices. Otherwise, they are considered as coordinate matrices
and the interpoint distances are calculated with \code{dist} and the method specified.
}
\examples{
n= 100

square = data.frame("x"=numeric(n),"y"=numeric(n))
square$x = runif(n,0,1)
square$y = runif(n,0,1)

disc = data.frame("x"=numeric(n),"y"=numeric(n))
radius = 0.5
r = radius*sqrt(runif(n,0,1))
theta = runif(n,0,1)*2*pi
disc$x= r*cos(theta)
disc$y= r*sin(theta)

DoD.stat(square, disc)
}
\references{
{
  \insertAllCited{}
}
}
