% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoD-Test.R
\name{DoD.test}
\alias{DoD.test}
\title{The (bootstrapped) Distribution-of-Distances (DoD) Test}
\usage{
DoD.test(
  input.asym,
  input.comp,
  beta = 0.01,
  p = 2,
  boot.m.out.of.n = dim(input.asym)[1],
  boot.rep = 1000,
  force.coordinates = F,
  method = "euclidean"
)
}
\arguments{
\item{input.asym}{either a n x d numeric coordinate matrix (resp. data frame) or a n x n distance matrix. The asymptotic distribution of the test statistic is approximated based on this input.}

\item{input.comp}{either a m x d numeric coordinate matrix (resp. data frame) or a m x m distance matrix.}

\item{beta}{the trimming parameter of the DoD-statistic. It can attain values in [0,1/2).}

\item{p}{the order of the DoD-statistic. By definition p>= 1.}

\item{boot.m.out.of.n}{the bootstrap sample size.}

\item{boot.rep}{the number of bootstrap repetitions.}

\item{force.coordinates}{boolean that indicates whether the input is forced to be interpreted as coordinate matrix.}

\item{method}{the distance measure to be used in case the input is considered as coordinate matrices. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}
}
\value{
The function returns a list containing the following components: \tabular{ll}{
\code{scaled.statistic}:  the value of the DoD-statistic (defined in \insertCite{weitkamp2020gromov;textual}{DoD}) scaled by mn/(m+n). \cr
\tab \cr
\code{p.value}: the p-value of the comparison.
}
}
\description{
Perform the (bootstrapped) DoD-test proposed in \insertCite{weitkamp2020gromov;textual}{DoD} for two samples of size n and m.
}
\details{
The DoD-Test is an asymptotic level-alpha test for the null hypothesis

\strong{H:}  \emph{"The distributions of (interpoint) distances of both inputs are equal"}
  
against the alternative that they are not. The test is based on the (trimmed) empirical Kantorovich distance of order p between the distribution of distances of the input
samples \code{input.asym} and \code{input.comp} (for more details see \insertCite{weitkamp2020gromov;textual}{DoD}).
 
If both inputs are quadratic, numeric matrices (and \code{force.coordinates}==F), they will be interpreted as distance matrices. Otherwise, they are considered as coordinate matrices
and the interpoint distances are calculated with \code{dist} and the method specified.

The limit distribution under the null hypothesis, which is required for testing,
is approximated by a resampling scheme (see Section 3 of \insertCite{weitkamp2020gromov;textual}{DoD}) based on the input \code{input.asym}.
}
\examples{
n= 100

square = data.frame("x"=numeric(n),"y"=numeric(n))
square$x = runif(n,0,1)
square$y = runif(n,0,1)

disc = data.frame("x"=numeric(n),"y"=numeric(n))
radius = 0.5
r = radius*sqrt(runif(n,0,1))
theta = runif(n,0,1)*2*pi
disc$x= r*cos(theta)
disc$y= r*sin(theta)

DoD.test(square, disc)
}
\references{
{
  \insertAllCited{}
}
}
